Depreate create/2 to avoid complex function guards
Writer.create/2 and Writer.create/3 for calls that instruct CouchDB to
generate an id compete with existing calls where clients have to provide
an id. Introducing the new name create_generate and deprecating the old
syntax will avoid confusion and clashes.
The two calls are also semantically different, so it makes sense to use
different names for them.