Document the reason for _id removal if _id is nil
This is an edge case and maybe raising an exception would be more
appropriate here but we currently prefer to remove the nil _id and
storing the document, thus returning a location for the new document.

It may be wiser to raise an exception here so this part is a candidate
for a breaking API change.